Separated some methods out of the HomeController. Fixed the Delete View

+++ Icarus Item Calculator/Icarus Item Calculator/Controllers/HomeController.cs
@@ -2,17 +2,22 @@ using Icarus_Item_Calculator.Models;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
+using Icarus_Item_Calculator.Services;
 
 namespace Icarus_Item_Calculator.Controllers
 {
     public class HomeController : Controller
     {
         private readonly ItemContext _context;
+        private readonly ItemServices _itemServices;
 
-        public HomeController(ItemContext context)
+
+        public HomeController(ItemContext context, ItemServices itemServices)
         {
             _context = context;
+            _itemServices = itemServices;
         }
+
 
         // GET: Home/Index
         public async Task<IActionResult> Index(string sortOrder, string searchString)
@@ -51,6 +56,7 @@ namespace Icarus_Item_Calculator.Controllers
             return View(await items.AsNoTracking().ToListAsync());
         }
 
+
         // GET: Home/Create
         public IActionResult Create()
         {
@@ -65,9 +71,11 @@ namespace Icarus_Item_Calculator.Controllers
             return View(item);
         }
 
+
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("Name,IsCraftable,IsCraftingInterface")] Item item, int[] selectedItems, Dictionary<double, double> quantities)
+        public async Task<IActionResult> Create([Bind("Name,IsCraftable,IsCraftingInterface")] Item item,
+                                                        int[] selectedItems, Dictionary<double, double> quantities)
         {
             if (ModelState.IsValid)
             {
@@ -102,6 +110,7 @@ namespace Icarus_Item_Calculator.Controllers
             return View(item);
         }
 
+
         // GET: Home/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
@@ -122,10 +131,12 @@ namespace Icarus_Item_Calculator.Controllers
             return View(item);
         }
 
+
         // POST: Home/Edit/5
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("ItemId,Name,IsCraftable,IsCraftingInterface")] Item item, int[] selectedItems, Dictionary<double, double> quantities)
+        public async Task<IActionResult> Edit(int id, [Bind("ItemId,Name,IsCraftable,IsCraftingInterface")] Item item,
+                                                int[] selectedItems, Dictionary<double, double> quantities)
         {
             if (id != item.ItemId)
             {
@@ -189,6 +200,11 @@ namespace Icarus_Item_Calculator.Controllers
             ViewBag.Items = new SelectList(_context.Items, "ItemId", "Name");
             return View(item);
         }
+        private bool ItemExists(int id)
+        {
+            return _context.Items.Any(e => e.ItemId == id);
+        }
+
 
         // GET: Home/Delete/5
         public async Task<IActionResult> Delete(int? id)
@@ -208,38 +224,24 @@ namespace Icarus_Item_Calculator.Controllers
             return View(item);
         }
 
+
         // POST: Home/Delete/5
         [HttpPost, ActionName("Delete")]
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var item = await _context.Items.FindAsync(id);
+            if (item == null)
+            {
+                return NotFound();
+            }
             _context.Items.Remove(item);
             await _context.SaveChangesAsync();
             return RedirectToAction(nameof(Index));
         }
 
-        private bool ItemExists(int id)
-        {
-            return _context.Items.Any(e => e.ItemId == id);
-        }
 
-        private async Task LoadNestedRecipes(Item item)
-        {
-            foreach (var recipeItem in item.Recipe)
-            {
-                if (recipeItem.Item != null)
-                {
-                    await _context.Entry(recipeItem.Item)
-                        .Collection(i => i.Recipe)
-                        .Query()
-                        .Include(r => r.Item)
-                        .LoadAsync();
 
-                    await LoadNestedRecipes(recipeItem.Item);
-                }
-            }
-        }
 
         public async Task<IActionResult> Details(int? id)
         {
@@ -281,73 +283,18 @@ namespace Icarus_Item_Calculator.Controllers
                 return NotFound();
             }
 
-            await LoadNestedRecipes(item);
-
-            var (recipeSteps, baseItemsTotal) = CalculateRecipeSteps(item, quantity);
+            await _itemServices.LoadNestedRecipesAsync(item);
+            var (recipeSteps, baseItemsTotal) = _itemServices.CalculateRecipeSteps(item, quantity);
 
             var model = new ItemWithStepsViewModel
             {
                 Item = item,
                 RecipeSteps = recipeSteps,
                 Quantity = quantity,
-                BaseItemsTotal = baseItemsTotal  // Add this to your view model
+                BaseItemsTotal = baseItemsTotal 
             };
 
             return View(model);
         }
-
-        private (List<RecipeStep>, Dictionary<string, double>) CalculateRecipeSteps(Item item, double quantity)
-        {
-            List<RecipeStep> steps = new List<RecipeStep>();
-            Dictionary<string, double> baseItemsTotal = new Dictionary<string, double>();
-            // Recursive function to break down the recipe
-            CalculateStepsRecursive(item, quantity, steps, new Dictionary<int, Dictionary<string, double>>(), baseItemsTotal);
-            return (steps, baseItemsTotal);
-        }
-        private void CalculateStepsRecursive(Item item, double quantity, List<RecipeStep> steps, Dictionary<int, Dictionary<string, double>> accumulatedIngredients, Dictionary<string, double> baseItemsTotal)
-        {
-            if (item == null) return;
-            // We only initialize for the current item because we want to track for this specific run
-            accumulatedIngredients[item.ItemId] = new Dictionary<string, double>();
-
-            steps.Add(new RecipeStep
-            {
-                ItemName = item.Name,
-                Quantity = quantity,
-                Ingredients = item.Recipe.Select(r =>
-                {
-                    // calculate total quantity needed for ingredient
-                    double totalQuantity = r.Quantity * quantity;   
-                    
-                    accumulatedIngredients[item.ItemId][r.Item.Name] = totalQuantity;
-
-                    // If it's a base item, add or update in baseItemsTotal
-                    if (r.Item.IsBaseItem)
-                    {
-                        if (baseItemsTotal.ContainsKey(r.Item.Name))
-                        {
-                            baseItemsTotal[r.Item.Name] += totalQuantity;
-                        }
-                        else
-                        {
-                            baseItemsTotal[r.Item.Name] = totalQuantity;
-                        }
-                    }
-
-                    return new IngredientStep
-                    {
-                        ItemName = r.Item.Name,
-                        Quantity = accumulatedIngredients[item.ItemId][r.Item.Name],
-                        IsBase = r.Item.IsBaseItem
-                    };
-                }).ToList()
-            });
-
-            foreach (var recipeItem in item.Recipe.Where(r => !r.Item.IsBaseItem))
-            {
-                CalculateStepsRecursive(recipeItem.Item, recipeItem.Quantity * quantity, steps, accumulatedIngredients, baseItemsTotal);
-            }
-        }
-
     }
 }
